Create LightMoveController.cs
Created a rough light move script

Casts a ray a gets collision point and then moves lighttracker to that possition overtime. the Spotlight rotates to lookAt Lighttracker

(Set lighttracker as an empty on the same y plane as the "Water".
place script on Spotlight and should work fine)

Delay is WIP and broken.