Added Boat Spawner Script
Not the boat controller just a spawner

Boat Spawner object has a vector3 setting on it, change that and the red bounding box is where the boat will spawn in.

while in play mode press space to spawn a new boat.

the boat that is spawned is a prefab under the Prefabs folder... add any scripts to that and it will update all boats.

also replaced lighthouse placeholder with the Lighthouse model